Refactor styles: Migrate CSS files to SCSS for improved maintainability and consistency
