+++ src/modules/auth/Login.tsx
@@ -24,7 +24,7 @@ import { useCurrentUser } from "../../context/user.context";
 import { useErrorHandler } from "../../hooks/error-handler";
 import { LoginForm, loginSchema } from "../../schemas/schemas";
 import { loginUser } from "../../services/user.service";
-import classes from "../../theme/modules/auth.module.css";
+import classes from "../../theme/modules/auth.module.scss";
 import PublicGuard from "../guards/PublicGuard";
 
 export default function Login() {
